requirements.txt
confluent_kafka==1.9.2
firebase_admin==6.0.1
pandas==1.3.5
plotly==5.11.0
pyspark==3.3.1
python-dotenv==0.21.0
streamlit==1.9.0
streamlit_autorefresh==0.0.1
